pyproject.toml
[tool.poetry]
name = "yandesk-api"
version = "0.1.0"
description = ""
authors = ["Shub Raj Lama <shuvraj1234@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
SQLAlchemy = "^1.4.40"
fastapi = {extras = ["full"], version = "^0.81.0"}
uvicorn = "^0.18.3"
gunicorn = "^20.1.0"
mysqlclient = "^2.1.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
